added function to do affine+lddmm and preprocessing in one step
