+++ ndreg/registerer.py
@@ -10,6 +10,7 @@ from matplotlib import pyplot as plt
 import shutil
 import util
 import preprocessor
+from itertools import product
 
 dimension = 3
 vectorComponentType = sitk.sitkFloat32
@@ -206,123 +207,6 @@ def imgMetamorphosisComposite(inImg, refImg, alphaList=0.02, betaList=0.05, scal
     return (compositeField, compositeInvField)
 
 
-def register_affine(atlas, img, learning_rate=1e-2, iters=200, min_step=1e-10, shrink_factors=[1], sigmas=[.150], use_mi=False, grad_tol=1e-6, verbose=False):
-    """
-    Performs affine registration between an atlas an an image given that they have the same spacing.
-    """
-    registration_method = sitk.ImageRegistrationMethod()
-
-    # Similarity metric settings.
-#     registration_method.SetMetricAsMeanSquares()
-    if use_mi: registration_method.SetMetricAsMattesMutualInformation(numberOfHistogramBins=128)
-    else: registration_method.SetMetricAsMeanSquares()
-    # registration_method.SetMetricSamplingStrategy(registration_method.RANDOM)
-    # registration_method.SetMetricSamplingPercentage(0.01)
-
-    registration_method.SetInterpolator(sitk.sitkBSpline)
-
-    # Optimizer settings.
-    registration_method.SetOptimizerAsRegularStepGradientDescent(learningRate=learning_rate,
-                                                                 minStep=min_step,
-    #                                                              estimateLearningRate=registration_method.EachIteration,
-                                                                 gradientMagnitudeTolerance=grad_tol,
-                                                                 numberOfIterations=iters)
-    registration_method.SetOptimizerScalesFromPhysicalShift()
-
-    # Setup for the multi-resolution framework.            
-    registration_method.SetShrinkFactorsPerLevel(shrinkFactors=shrink_factors)
-    registration_method.SetSmoothingSigmasPerLevel(smoothingSigmas=sigmas)
-    registration_method.SmoothingSigmasAreSpecifiedInPhysicalUnitsOn()
-
-    # initial transform
-    initial_transform = sitk.AffineTransform(atlas.GetDimension())
-    length = np.array(atlas.GetSize())*np.array(atlas.GetSpacing())
-    initial_transform.SetCenter(length/2.0)
-
-    # Don't optimize in-place, we would possibly like to run this cell multiple times.
-    registration_method.SetInitialTransform(initial_transform)
-
-    # Connect all of the observers so that we can perform plotting during registration.
-    if verbose:
-        registration_method.AddCommand(sitk.sitkStartEvent, start_plot)
-        registration_method.AddCommand(sitk.sitkEndEvent, end_plot)
-        registration_method.AddCommand(sitk.sitkMultiResolutionIterationEvent, update_multires_iterations) 
-        registration_method.AddCommand(sitk.sitkIterationEvent, lambda: plot_values(registration_method))
-
-    final_transform = registration_method.Execute(sitk.Cast(img, sitk.sitkFloat32),
-                                                  sitk.Cast(atlas, sitk.sitkFloat32))
-    return final_transform
-
-def register_lddmm(affine_img, target_img, alpha_list=0.05, scale_list=[0.0625, 0.125, 0.25, 0.5, 1.0], 
-                   epsilon_list=1e-4, min_epsilon_list=1e-10, sigma=0.1, use_mi=False, iterations=200, inMask=None,
-                   refMask=None, verbose=True, out_dir=''):
-    if sigma == None:
-        sigma = (0.1/target_img.GetNumberOfPixels())
-
-    (field, invField) = imgMetamorphosisComposite(affine_img, target_img,
-                                                                                                alphaList=alpha_list,
-                                                                                                scaleList=scale_list,
-                                                                                                epsilonList=epsilon_list,
-                                                                                                minEpsilonList=min_epsilon_list,
-                                                                                                sigma=sigma,
-                                                                                                useMI=use_mi,
-                                                                                                inMask=inMask,
-                                                                                                refMask=refMask,
-                                                                                                iterations=iterations, 
-                                                                                                verbose=verbose,
-                                                                                                outDirPath=out_dir)
-
-    source_lddmm = imgApplyField(affine_img, field, 
-                                            size=target_img.GetSize(), 
-                                            spacing=target_img.GetSpacing())
-    return source_lddmm, field, invField
-
-def register_rigid(atlas, img, learning_rate=1e-2, iters=200, min_step=1e-10, shrink_factors=[1], sigmas=[.150], use_mi=False, grad_tol=1e-6, verbose=False):
-    """
-    Performs affine registration between an atlas an an image given that they have the same spacing.
-    """
-    registration_method = sitk.ImageRegistrationMethod()
-
-    # Similarity metric settings.
-#     registration_method.SetMetricAsMeanSquares()
-    if use_mi: registration_method.SetMetricAsMattesMutualInformation(numberOfHistogramBins=128)
-    else: registration_method.SetMetricAsMeanSquares()
-    # registration_method.SetMetricSamplingStrategy(registration_method.RANDOM)
-    # registration_method.SetMetricSamplingPercentage(0.01)
-
-    registration_method.SetInterpolator(sitk.sitkBSpline)
-
-    # Optimizer settings.
-    registration_method.SetOptimizerAsRegularStepGradientDescent(learningRate=learning_rate,
-                                                                 minStep=min_step,
-    #                                                              estimateLearningRate=registration_method.EachIteration,
-                                                                 gradientMagnitudeTolerance=grad_tol,
-                                                                 numberOfIterations=iters)
-    registration_method.SetOptimizerScalesFromPhysicalShift()
-
-    # Setup for the multi-resolution framework.            
-    registration_method.SetShrinkFactorsPerLevel(shrinkFactors=shrink_factors)
-    registration_method.SetSmoothingSigmasPerLevel(smoothingSigmas=sigmas)
-    registration_method.SmoothingSigmasAreSpecifiedInPhysicalUnitsOn()
-
-    # initial transform
-    initial_transform = sitk.VersorRigid3DTransform()
-    length = np.array(atlas.GetSize())*np.array(atlas.GetSpacing())
-    initial_transform.SetCenter(length/2.0)
-
-    # Don't optimize in-place, we would possibly like to run this cell multiple times.
-    registration_method.SetInitialTransform(initial_transform)
-
-    # Connect all of the observers so that we can perform plotting during registration.
-    if verbose:
-        registration_method.AddCommand(sitk.sitkStartEvent, start_plot)
-        registration_method.AddCommand(sitk.sitkEndEvent, end_plot)
-        registration_method.AddCommand(sitk.sitkMultiResolutionIterationEvent, update_multires_iterations) 
-        registration_method.AddCommand(sitk.sitkIterationEvent, lambda: plot_values(registration_method))
-
-    final_transform = registration_method.Execute(sitk.Cast(img, sitk.sitkFloat32),
-                                                  sitk.Cast(atlas, sitk.sitkFloat32))
-    return final_transform
 
 def resample(image, transform, ref_img, default_value=0.0):
     # Output image Origin, Spacing, Size, Direction are taken from the reference
@@ -469,81 +353,6 @@ def createTmpRegistration(inMask=None, refMask=None,
         tmpRegistration.SetMetricFixedMask(refMask)
 
     return tmpRegistration
-
-# Utility functions for plotting
-
-from ipywidgets import interact, fixed
-from IPython.display import clear_output
-
-# Callback invoked by the interact IPython method for scrolling through the image stacks of
-# the two images (moving and fixed).
-def display_images(fixed_image_z, moving_image_z, fixed_npa, moving_npa):
-    # Create a figure with two subplots and the specified size.
-    plt.subplots(1,2,figsize=(10,8))
-    
-    # Draw the fixed image in the first subplot.
-    plt.subplot(1,2,1)
-    plt.imshow(fixed_npa[fixed_image_z,:,:],cmap=plt.cm.Greys_r)
-    plt.title('fixed image')
-    plt.axis('off')
-    
-    # Draw the moving image in the second subplot.
-    plt.subplot(1,2,2)
-    plt.imshow(moving_npa[moving_image_z,:,:],cmap=plt.cm.Greys_r)
-    plt.title('moving image')
-    plt.axis('off')
-    
-    plt.show()
-
-# Callback invoked by the IPython interact method for scrolling and modifying the alpha blending
-# of an image stack of two images that occupy the same physical space. 
-def display_slices_with_alpha(fixed, moving, alpha, vmax):
-    img = (1.0 - alpha)*fixed + alpha*moving
-    plt.imshow(sitk.GetArrayViewFromImage(img),cmap=plt.cm.Greys_r, vmax=vmax)
-    plt.axis('off')
-    plt.show()
-
-def display_images_with_alpha(slice_num, alpha, fixed, moving, axis=2, vmax=1000):
-    if axis == 0: display_slices_with_alpha(fixed[slice_num,:,:], moving[slice_num,:,:], alpha, vmax)
-    elif axis == 1: display_slices_with_alpha(fixed[:,slice_num,:], moving[:,slice_num,:], alpha, vmax)
-    else: display_slices_with_alpha(fixed[:,:,slice_num], moving[:,:,slice_num], alpha, vmax)
-    
-# Callback invoked when the StartEvent happens, sets up our new data.
-def start_plot():
-    global metric_values, multires_iterations
-    
-    metric_values = []
-    multires_iterations = []
-
-# Callback invoked when the EndEvent happens, do cleanup of data and figure.
-def end_plot():
-    global metric_values, multires_iterations
-    
-    del metric_values
-    del multires_iterations
-    # Close figure, we don't want to get a duplicate of the plot latter on.
-    plt.close()
-
-# Callback invoked when the IterationEvent happens, update our data and display new figure.    
-def plot_values(registration_method):
-    global metric_values, multires_iterations
-    
-    metric_values.append(registration_method.GetMetricValue())                                       
-    # Clear the output area (wait=True, to reduce flickering), and plot current data
-    clear_output(wait=True)
-    # Plot the similarity metric values
-    plt.plot(metric_values, 'r')
-    plt.plot(multires_iterations, [metric_values[index] for index in multires_iterations], 'b*')
-    plt.xlabel('Iteration Number',fontsize=12)
-    plt.ylabel('Metric Value',fontsize=12)
-    plt.show()
-    
-# Callback invoked when the sitkMultiResolutionIterationEvent happens, update the index into the 
-# metric_values list. 
-def update_multires_iterations():
-    global metric_values, multires_iterations
-    multires_iterations.append(len(metric_values))  
-    
 
 def imgNorm(img):
     """
